// client/api/__generated__/types/visit.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * API
 * API description
 * OpenAPI spec version: 0.1
 */
import type { VisitStatus } from './visitStatus';
import type { Patient } from './patient';

export interface Visit {
  id: string;
  visitDate: string;
  diagnosis: string;
  treatment: string;
  status: VisitStatus;
  /** @nullable */
  notes?: string | null;
  patientId: string;
  patient?: Patient;
  createdAt: string;
  updatedAt: string;
}